// brokerchain-1.0.0/src/library/storage-simple-coll/collection-get/type.ts
// auto generated by dev/system

import { Logger } from "../../../myutils/logger.js";

export interface Input {
    namespace: string;
    key: string;
}

export interface OutputOk {
    items: { id: string; [key: string]: any }[];
}

export interface OutputFail {}

export interface Callback<R> {
    empty: () => R;
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}

export function check_input<R>(plog: Logger, v: any, cb: { ok: () => R; fail: (err: Error) => R }): R {
    const log = plog.sub("check_input");
    log.variable("v", v);
    try {
        log.println("v must be object");
        if (typeof v !== "object" || v === null) {
            throw new Error("v is not object");
        }

        Object.keys(v).forEach((field) => {
            if (field === "namespace") return;
            if (field === "key") return;
            throw new Error("v contains unknown field: " + field);
        });

        log.println("v.namespace must be string");
        if (typeof v.namespace !== "string") {
            throw new Error("v.namespace is not string");
        }

        log.println("v.key must be string");
        if (typeof v.key !== "string") {
            throw new Error("v.key is not string");
        }
    } catch (err) {
        log.error(err);
        return cb.fail(err);
    }
    // nothing wrong
    log.ok();
    return cb.ok();
}

// JSON stringify value before copy to handle some specific problem
// eg. Date Object probelm
export function copy_output_ok(v: any): OutputOk {
    if (v !== undefined) {
        const v_cloned = JSON.parse(JSON.stringify(v));
        return _copy_output_ok(v_cloned);
    } else {
        return _copy_output_ok(v);
    }
}

export function _copy_output_ok(v: any): OutputOk {
    if (typeof v === "object" && v !== null) {
        const obj = {
            items: copy_items(v.items)
        };
        return obj;
    } else {
        return { items: [] };
    }

    function copy_items(v: any): { id: string; [key: string]: any }[] {
        return Array.isArray(v) ? v.map(copy_item) : [];

        function copy_item(v: any): { id: string; [key: string]: any } {
            if (typeof v === "object" && v !== null) {
                const obj = {
                    id: copy_id(v.id)
                };
                return apply_dynamic_fields(obj);
            } else {
                return { id: "" };
            }

            function copy_id(v: any): string {
                return typeof v === "string" ? v : "";
            }

            function apply_dynamic_fields(dest: any) {
                // FIXME this implementaion is very limited...
                // maybe we should never copy the dynamic fields?
                Object.keys(v).forEach((name) => {
                    dest[name] = copy_dynamic_field(v[name]);
                });
                return dest;

                function copy_dynamic_field(v: any): any {
                    // the value type here is any
                    if (v === undefined || v === null) {
                        return v;
                    } else if (typeof v === "boolean") {
                        return v;
                    } else if (typeof v === "number") {
                        return v;
                    } else if (typeof v === "string") {
                        return v;
                    } else {
                        return JSON.parse(JSON.stringify(v));
                    }
                }
            }
        }
    }
}

// backward compatible purpose
export type Output = OutputOk;
export const copy_output = copy_output_ok;